Some beauty fixes in the code; HP and health bar calculation only for actors
